requirements.txt
streamlit>=1.28.0
openai>=1.3.0
python-dotenv>=1.0.0
audio-recorder-streamlit>=0.0.8
pydub>=0.25.1
requests>=2.31.0
pandas>=2.0.0
numpy>=1.24.0
plotly>=5.17.0
streamlit-chat>=0.1.1
streamlit-audiorecorder>=0.0.5